Add --watch mode to fingerprint command
