+++ package.json
@@ -36,6 +36,7 @@
   ],
   "dependencies": {
     "async": "^3.2.0",
+    "chokidar": "^3.5.1",
     "core-js": "^3.6.5",
     "crypto-js": "^4.0.0",
     "d3": "^5.16.0",
@@ -43,6 +44,7 @@
     "dagre": "^0.8.5",
     "deepmerge": "^4.2.2",
     "diff": "^5.0.0",
+    "glob": "^7.1.6",
     "js-yaml": "^4.0.0",
     "oboe": "^2.1.5",
     "sql-formatter": "^2.3.3",
